We will work on subtitles. First when user clicks on subtitle icon, we will show a dialog showing all the directories of device storage and user will select the subtitle file from device storage and after clicking on OK the subtitle will start showing with the video. We will use an android (file picker) library for fetching all the files and directories of your phone and after selecting subtitle file, video will start showing subtitles, subtitle file extension should be in .srt.
Run app, open the installed app and open a folder > video file with subtitle. Expand list and click on the subtitle icon. The alert dialog contains all the files and directories of the device storage. You have to open the folder that contains the subtitle file with .srt extension. Select the file - user can select only one file at a time - and click on OK. The subtitles are playing fine.

SUBTITLES ON INCREASING SPEED
If you increase the speed, the subtitles are showing with video accordingly.

For now we have worked on only .srt extension of subtitle file. We later shall make it compatible for other extensions like .vtt, .ass, .ssa and some other extensions.