To fix the issue where the toast shows "Left Swipe" when swiping on the right side, and vice versa, you need to adjust the logic in your code where you determine whether the swipe occurred on the left or right side of the screen. Currently, it seems like there might be an error in the logic used to determine the side of the swipe.
With this correction, the code will correctly identify the side of the swipe, and the toast messages should display "Right Swipe" when swiping on the right side of the screen and "Left Swipe" when swiping on the left side.

By using else instead of else if, we ensure that if the touch event is not detected on the left side of the screen, it's considered to be on the right side. This correction ensures that the code accurately identifies the side of the swipe, fixing the issue where the toast message incorrectly showed "Left Swipe" when swiping on the right side of the screen.